+++ src/types.ts
@@ -1,13 +1,15 @@
 import { DataQuery, DataSourceJsonData } from '@grafana/data';
 
+// keep same as the loki data source
 export interface MyQuery extends DataQuery {
-  expr: string; // keep same as the loki data source
+  expr: string;
+  maxLines?: number;
 }
 
 /**
  * These are options configured for each DataSource instance
  */
 export interface MyDataSourceOptions extends DataSourceJsonData {
-  promDataSourceUid: string;
-  lokiDataSourceUid: string;
+  promDataSourceName: string;
+  lokiDataSourceName: string;
 }