+++ CHANGELOG.md
@@ -1,5 +1,50 @@
 # Changelog
 
-## 1.0.0 (Unreleased)
+## 0.0.9
 
-Initial release.
+[PR #7](https://github.com/baurine/customized-loki-ds-plugin/pull/7)
+
+- Fix the plugin validator warnings, prepare for publishing to grafana
+
+## 0.0.8
+
+[PR #5](https://github.com/baurine/customized-loki-ds-plugin/pull/5)
+
+- Refine search expression
+- Refine Help page
+
+## 0.0.7
+
+[PR #4](https://github.com/baurine/customized-loki-ds-plugin/pull/4)
+
+- Replace most `=~` to `=`, support normal and regex search string both
+- Refine Help page
+
+## 0.0.6
+
+- Support line limit option
+- Refine Help page
+- Adjust the order of the options
+
+## 0.0.5
+
+- Support select multiple LogType (aka containers)
+- Hide filter label if there is no filter
+- Adjust layout
+- Refine tooltip
+
+## 0.0.4
+
+- Test github action
+
+## 0.0.3
+
+- Auto publish release
+
+## 0.0.2
+
+- Change promDataSourceUid/lokiDataSourceUid to promDataSourceName/lokiDataSourceName
+
+## 0.0.1
+
+First release